Add option to configure accepted answers background color
The current background color `rgba(0, 40, 16, .5)` is can be difficult to look at. Adding an option to be able to configure this so it can be set to anything else, such as `rgba(36, 36, 36, .5)` to match the rest of the answers. 

The green border isn't changed as this gives a nice accent so you know you're still looking at the accepted answer. 
The alpha channel is still fixed to .5, but maybe this should be changed to not be fixed?