suppress inline animation
this animation is inline HTML and this is a quick fix that breaks the 
animation but styles the flash of whites.

may try support animation if anyone has a good reason.

Note this commit doesnt seem to break page loading.

See https://stackoverflow.com/posts/59450319/timeline#comment_105942525

click on the days text

disable style reload page and test to see animation that this commit 
removes.